Addressing CRAN review feedback:
Please add \value to .Rd files regarding exported methods and explain
the functions results in the documentation. Please write about the
structure of the output (class) and also what the output means. (If a
function does not return a value, please document that too, e.g.
\value{No return value, called for side effects} or similar)
  Missing Rd-tags:
      futureverse_packages.Rd: \value
      futureverse_update.Rd: \value

Please write references in the description of the DESCRIPTION file in
the form
authors (year) <doi:...>
authors (year, ISBN:...)
or if those are not available: authors (year) <https:...>
with no space after 'doi:', 'https:' and angle brackets for
auto-linking. (If you want to add a title as well please put it in
quotes: "Title")
So please add auhtors and (year) as well.